Logical Gates and loading.
All logical gates are written and going on with the loading part.
A problem arised: the map keeps the wrong pointers because the vector, while
increasing in size, copies the original items and the pointers are not up to
date with the real elements.
A possible solution is the introducion of smart pointers.
(๑ʘ∆ʘ๑)