Logical gates structure and done with the loading.
The pointers problem is solved using vectors of pointers: in this way when a
vector resizes itself, it copies the pointers and not the objects, so every
reference are going to be correct.
The Circuit distructor need to be implemented, so every item created will be
destroyed at the end of the problem.
All gates now has a .h and .cpp file, more clean in this way.
Added the DEBUG flag at the beginning of every .cpp file: if it's defined some
debug printing will show on the stdout, nothing will happen otherwise.
Loading from a file, or a series of file, complete and correct.
Next step is evaluating the circuit.
`(๑ △ ๑)`*