Solving and Solution
The circuit is now correctly computed (only for NAND gates, I have to fix a
problem with the NOT operator), the solutions stored and compared to each other
to see problems with the failing gates.
Importante information are written on stdout, they could be reworked for a
better rappresentation.
Next step: fix all gates, fix a problem when showing only the outputs values in
circuit, creating the output files for the second part of the problem.
|_・)