Logical Gates fix, printing improvements and fixes.
Fixed every logical gate that had the NOT operator. Every gate should be tested,
the only one that has been tested is the NAND gate.
Fixed the printing output value issue and improved the final user printings.
Added the possibility to pass the input configuration and failure gates files as
argument to the program.
Next step: creating the output file for the second part and start dealing with
that.
(/¯◡ ‿ ◡)/¯ ~ ┻━┻