Doc and scripts
Wrote something in the documentation: sort of introduction for the simulation
part.
Wrote scripts to create instances for the known circuits, solve one circuit and
solve every circuit.
Created the result file for the simulation.
Found a bug in the cone computation, when an Output Terminal is directly
connected to an Input Terminal: to be fixed.
Other than this bug, all the others instances seem to work correctly.
(^ᴥ^)