Diagnoses_one_choice
Created the four vectors to easily access every gate classified by their
qualitative output value.
Changed the cone_collection structure from a vector to a map: now I know also
whose cone is which, using the string gate name as key of the map.
Diagnoses_one_choice done: the most complex part was creating the current choice
combination. I totally nailed it! I did it in a similar way as done before for
the ok subset computation, but the numbers are represented depending on the
number of ko gates.
GateCone set operations changed: they no longer operates on the item itself, but
creates a copy and returns the result in it. In this way it's easier to handle
the resulting cones outside the class.
Everything seem to be fine, checked with the slide example and with a few other
just changing the gate values to test the combination procedure.
The last part is to integrate the MHS computation.
I also have to check how to properly do different types of diagnoses, the one
implemented by now is the full masking diagnoses.
( ͠° ͟ʖ ͡°)