Bug, times, scripts
Fixed a bug in the Simulation part: the cones were not correctly updated in the
LogicalGate dued to the changes made to how the join in GateCone are done.
Added times indicator underlying how much time is spent calculating the MHS in
comparison to the rest of the problem.
Updated the solve_circuit script to do also the Diagnostic: it does the
requested type of diagnoses, given as argument, or ALL_DIAGNOSES if that's
missing.
Run some tests and everything seem to work, MHS requires a freaking lot of time.
Next step: test and documentation.
~(>_<~)