Scripts and doc.
Modified scripts so they are more flexible: now it is possible to choose how
many times every circuit has to be solved and which type of diagnoses is
requested.
Written all the Diagnostic part, maybe a re-read is needed.
Next step: create and add the UML Class diagram of the Diagnostic part and write
the testing part.
( ب_ب )