Fixed a bug in combinations
Fixed a bug in both ok_subset and combinations_choice generation: when the base
sets grow in number, the combinations grow exponentially and an int wasn't
enough to store the result of all the possible combinations or to indicate which
combination we were computing. Changed to a double and changed the rest operator
(%) with the fmod function.
(҂⌣̀_⌣́)